mode list entry to file
